requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
beautifulsoup4>=4.12.0
duckdb==0.10.3
feedparser==6.0.11
fastapi==0.128.0
newspaper3k>=0.2.8
readability-lxml>=0.8.0
h11==0.16.0
idna==3.11
python-dotenv>=1.0.0
pandas==2.2.3
pydantic==2.12.5
pydantic_core==2.41.5
PyYAML==6.0.2
pyarrow==16.1.0
requests==2.32.5
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
